+++ apps/c/airfoil/SP/airfoil_kernels.cpp
@@ -10,13 +10,13 @@ void __syncthreads(){}
                                
 // global constants            
                                
-extern float gam;              
-extern float gm1;              
-extern float cfl;              
-extern float eps;              
-extern float mach;             
-extern float alpha;            
-extern float qinf[4];          
+extern double gam;              
+extern double gm1;              
+extern double cfl;              
+extern double eps;              
+extern double mach;             
+extern double alpha;            
+extern double qinf[4];          
                                
 // user kernel files
