+++ apps/c/airfoil/SP/save_soln.h
@@ -1,3 +1,3 @@
-inline void save_soln(float *q, float *qold){
+inline void save_soln(double *q, double *qold){
   for (int n=0; n<4; n++) qold[n] = q[n];
 }
