+++ apps/c/jac1/update.h
@@ -1,4 +1,4 @@
-inline void update(float *r, float *du, float *u, float *u_sum, float *u_max){
+inline void update(double *r, double *du, double *u, double *u_sum, double *u_max){
   *u += *du + alpha * (*r);
   *du = 0.0f;
   *u_sum += (*u)*(*u);
